+++ framework/model/src/main/java/is/codion/framework/model/EntityTableConditionModel.java
@@ -71,22 +71,9 @@ public interface EntityTableConditionModel extends TableConditionModel<Attribute
 	ForeignKeyConditionModel get(ForeignKey foreignKey);
 
 	/**
-	 * Controls the additional WHERE condition. The condition supplier may return null in case of no condition.
-	 * Note that in order for the {@link #changed()} {@link is.codion.common.reactive.observer.Observer} to indicate
-	 * a changed condition, the additional condition must be set via {@link AdditionalCondition#set(Object)},
-	 * changing the return value of the underlying {@link Supplier} instance does not trigger a changed condition.
-	 * @return the {@link AdditionalCondition} instance controlling the additional WHERE condition
+	 * @return the {@link AdditionalConditions} instance, for managing additional conditions
 	 */
-	AdditionalCondition where();
-
-	/**
-	 * Controls the additional WHERE condition. The condition supplier may return null in case of no condition.
-	 * Note that in order for the {@link #changed()} {@link is.codion.common.reactive.observer.Observer} to indicate
-	 * a changed condition, the additional condition must be set via {@link AdditionalCondition#set(Object)},
-	 * changing the return value of the underlying {@link Supplier} instance does not trigger a changed condition.
-	 * @return the {@link AdditionalCondition} instance controlling the additional HAVING condition
-	 */
-	AdditionalCondition having();
+	AdditionalConditions additional();
 
 	/**
 	 * @return the {@link Modified} instance
@@ -106,9 +93,33 @@ public interface EntityTableConditionModel extends TableConditionModel<Attribute
 	}
 
 	/**
-	 * Specifies an additional condition supplier.
+	 * Manages the additional WHERE and HAVING conditions.
 	 */
-	interface AdditionalCondition extends Value<Supplier<Condition>> {
+	interface AdditionalConditions {
+
+		/**
+		 * Controls the additional WHERE condition. The condition supplier may return null in case of no condition.
+		 * Note that in order for the {@link #changed()} {@link is.codion.common.reactive.observer.Observer} to indicate
+		 * a changed condition, the additional condition must be set via {@link ConditionValue#set(Object)},
+		 * changing the return value of the underlying {@link Supplier} instance does not trigger a changed condition.
+		 * @return the {@link ConditionValue} instance controlling the additional WHERE condition
+		 */
+		ConditionValue where();
+
+		/**
+		 * Controls the additional WHERE condition. The condition supplier may return null in case of no condition.
+		 * Note that in order for the {@link #changed()} {@link is.codion.common.reactive.observer.Observer} to indicate
+		 * a changed condition, the additional condition must be set via {@link ConditionValue#set(Object)},
+		 * changing the return value of the underlying {@link Supplier} instance does not trigger a changed condition.
+		 * @return the {@link ConditionValue} instance controlling the additional HAVING condition
+		 */
+		ConditionValue having();
+	}
+
+	/**
+	 * Manages an additional condition supplier.
+	 */
+	interface ConditionValue extends Value<Supplier<Condition>> {
 
 		/**
 		 * Default {@link Conjunction#AND}.